+++ activesupport/lib/active_support/cache/memory_store.rb
@@ -30,6 +30,13 @@ module ActiveSupport
         @pruning = false
       end
 
+      # Advertise that this cache store can be used
+      # with "recyclable cache keys" otherwise known
+      # as cache versioning.
+      def supports_in_cache_versioning?
+        true
+      end
+
       # Delete all data stored in a given cache store.
       def clear(options = nil)
         synchronize do